+++ tests_integration/test_connection.py
@@ -62,3 +62,15 @@ def test_connection_not_existing_schema(
             f'Could not use schema "{schema.upper()}". Object does not exist'
             in result.output
         )
+
+
+@pytest.mark.parametrize("report_dir", [".", "custom/report/dir"])
+@pytest.mark.integration
+def test_connection_diagnostic_report(runner, report_dir, temporary_working_directory):
+    from pathlib import Path
+
+    command = ["connection", "test", "--enable-diag", "--diag-log-path", report_dir]
+    result = runner.invoke_with_connection(command)
+    assert result.exit_code == 0, result.output
+    expected_report_location = Path(report_dir) / "SnowflakeConnectionTestReport.txt"
+    assert expected_report_location.exists(), result.output